Use SSL_CERT_FILE env var if set in test builds
When the `_test-util` feature is enabled, check to see if
`SSL_CERT_FILE` is set and if that's the case use the native roots
rustls connector configuration, which expects that variable to point at
a pem file.

This allows a self signed certificate to be used for testing purposes,
without altering the validation path, but retains webpki roots for
normal builds or if the environment variable is not set (i.e. the
platform supplied root certificates are not used for test builds, the
environment variable must be set for this to take effect).