Introduce Prober for directory opt-in
RFC 9458 (OHTTP) specifies a 1:1 mapping between relays and gateways in
order to preclude abuse. That presents an issue for payjoin UX, the
receiver chooses the directory, which is its own OHTTP gateway, but then
the sender must know a relay that will forward requests to that
directory.

In order to eliminate the need for maintaining up to date relay
configurations, this commit adds a mechanism that allows OHTTP gateways
to signal to relays that they will accept OHTTP requests for a specified
purpose (i.e. BIP77), even if the relay is not configured to allow that
specific gateway a priori.

The new opt-in mechanism is extends RFC 9540: when querying the
`/.well-known/ohttp-gateway` endpoint with query string
`?allowed_purposes`, if the gateway responds with an
`application/x-ohttp-allowed-purposes` response whose body is a list of
bytestrings encoded in the same format as TLS ALPN (U16 big endian
element count list, with U8 length prefixed bytestring), and this
response contains the magic string
`BIP77 454403bb-9f7b-4385-b31f-acd2dae20b7e` as one of the elements,
this indicates that the relay may forward BIP77 OHTTP requests to it.

The relay now expects POST requests to include the gateway in the
request path. Only the scheme and authority are to be included (same as
GatewayUri struct), the RFC 9540 path on the gateway's host is implied.

When a gateway does not opt-in, this is treated as 404 Not Found.

When the prober is over capacity, the relay will respond with 503
Service Unavailable whenever requests to unknown gateways are made.
Retry-After will be set for these responses according to the TTL of the
soonest expiring gateway status.