Refactor GatewayUri
Use a non optional Scheme & Authority for representing the internal
state, instead of a Uri which may be relative, and may contain a path.

Authority is normalized to have an explicit port, as before.

Resolution of authority to a socket addr is now async, to avoid blocking
while waiting for host resolution, in anticipation of gateway opt-in
allowing adversary chosen hostnames to be resolved which may be a DoS
concern.

Closes #51.