Don't cargo update in MSRV CI job
Since not all dependencies have precise semvers constraints, a blanket
cargo update can break MSRV.

This means that the repository can bitrot, with CI jobs failing in ways
not related to changes to the codebase.

This is a workaround, see also #43 for description of the proper way to
fix this.