Utilize gateway prober in relay
The relay now expects POST requests to include the gateway in the
request path. Only the scheme and authority are to be included (same as
GatewayUri struct), the RFC 9540 path on the gateway's host is implied.

When a gateway does not opt-in, this is treated as 404 Not Found.

When the prober is over capacity, the relay will respond with 503
Service Unavailable whenever requests to unknown gateways are made.
Retry-After will be set for these responses according to the TTL of the
soonest expiring gateway status.