Test with lock file, don't force update
Running cargo update at msrv lets you know when an additional pin
needs to be made if running without a lock file. But we can just commit
a lock file known to work with MSRV to solve our problems.

However, discovering the pins to produce the lock file is tricky, so we
still want to have a script that documents the pins required for MSRV.