Make HttpClient use a BoxBody, correct headers
In order to allow the prober to share the same client as the proxy, the
Request<Incoming> type needs to be boxed, as even an empty Incoming body
type can't be created outside of the hyper crate.

When converting the request, two changes in behavior were also
implemented:

No longer set Host header to 0.0.0.0. Instead, let the client set it
from the request URL's authority.

Also specify Content-Type as `message/ohttp-req`, this was previously
omitted from the forwarded request.