Update nix flake to use nixos-unstable
This commit updates the nix flake to use nixos-unstable as our msrv is
now 1.85.0, with some build deps now requiring this higher version of
rust we were getting comiler errors when trying to build on this flake.

This change also keeps this flake more inline with the main rust-payjoin
repo which also uses unstable for the flake.